Added powershell file for Azure Runbook
The intention is to scale SQL Azure in Azure Runbook using Az Powershell. Do note that Get-AutomationConnection is used and not Get-AzAutomationConnection which was throwing an "object not set" error. "AzureRunAsConnection" is set up for the Automation Account and it must have access to the target resource. Value of $Edition must lead to value of #Capacity.